# openapi.yaml
openapi: 3.0.1
info:
  title: Python Execution Plugin Enhanced
  description: A plugin that allows the user to execute Python code, create Python files, install Python packages, and run unit tests using ChatGPT.
  version: 'v1'
servers:
  - url: http://localhost:3333
paths:
  /execute:
    post:
      operationId: executeCode
      summary: Execute Python code
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /create:
    post:
      operationId: createFile
      summary: Create a Python file
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                code:
                  type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /install:
    post:
      operationId: installPackage
      summary: Install a Python package
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                package:
                  type: string
                package_name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string

  /run-unit-tests:
    post:
      operationId: runUnitTests
      summary: Run unit tests in a Python file
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string